feat: refactor controllers and services for admin-specific functionality and user enrollment

+++ src/main/java/vn/fu_ohayo/service/JwtService.java
@@ -7,7 +7,7 @@ import vn.fu_ohayo.enums.TokenType;
 import java.util.Collection;
 
 public interface JwtService {
-    String generateAccessToken(int userId, String email, Collection<? extends GrantedAuthority> authorities);
-    String generateRefreshToken(int userId, String email, Collection<? extends GrantedAuthority> authorities);
+    String generateAccessToken(long userId, String email, Collection<? extends GrantedAuthority> authorities);
+    String generateRefreshToken(long userId, String email, Collection<? extends GrantedAuthority> authorities);
     ExtractTokenResponse extractUserInformation(String token, TokenType type);
 }
